add check at top of cleanup for plist files. added documentation to script.sh.

+++ script.sh
@@ -3,35 +3,64 @@
 # Convert minutes to seconds, then create a one-time cron that
 # simply calls up a sticky growlnotify with your reminder. 
 
+GROWL=/usr/local/bin/growlnotify
+
+
+### Cleanup Script Begins ###
 
 if [[ $1 == "cleanup" ]]; then
 
-	echo "Cleaning up expired reminders..."
-	ACTIVE_REMINDERS=($(launchctl list | grep com.approductive.remindersapp | awk '{print $3}'))
+	# get the list of plist files this user has created
 	REMINDER_PLISTS=($(command ls ~/Library/LaunchAgents/com.approductive.remindersapp.*))
+	
+	# if there are plists, let's proceed to see if any are expired
+	if [[ -n $REMINDER_PLISTS ]]; then
 
-	total_plists=${#REMINDER_PLISTS[*]}
-	for (( i=0; i<=$(( $total_plists -1 )); i++ ))
-	do
-		trimmed_plist=$(basename ${REMINDER_PLISTS[$i]} | sed 's#.plist##g')
-		is_active=$(echo "${ACTIVE_REMINDERS[@]}" | grep -o "$trimmed_plist")
+		echo "Cleaning up expired reminders..."
 		
-		if [[ -z $is_active ]]; then
-			rm ${REMINDER_PLISTS[$i]}
-		fi
-	done
-	exit
+		# get list of active reminders
+		ACTIVE_REMINDERS=($(launchctl list | grep com.approductive.remindersapp | awk '{print $3}'))
+		
+		# get total count of plist files to iterate	
+		total_plists=${#REMINDER_PLISTS[*]}
+		for (( i=0; i<=$(( $total_plists -1 )); i++ ))
+		do
+			# remove the .plist so it will match the launchctl output for comparison
+			trimmed_plist=$(basename ${REMINDER_PLISTS[$i]} | sed 's#.plist##g')
+			# compare existing plist file against active launchctl list
+			is_active=$(echo "${ACTIVE_REMINDERS[@]}" | grep -o "$trimmed_plist")
+			
+			# if the plist is NOT active, remove the file
+			if [[ -z $is_active ]]; then
+				rm ${REMINDER_PLISTS[$i]}
+			fi
+		done
+	fi
+
+	# exit this script if there were plist files we cleaned up or not
+	exit 0
+
 fi
 
+### Cleanup Script Ends ###
+
+
+### Create Reminder Begins ###
+# Format: remindme 1 the_reminder
+# Format: remindme $1 $2
+
+# Convert minutes to epoch
 MINUTES=$1
 TIMESTAMP=$(command date +%s)
 TIMER=$(($1 * 60))
-GROWL=/usr/local/bin/growlnotify
+
+# Capture all remaining arguments as $REMINDER
 shift
 REMINDER=$*
 echo "$MINUTES minute reminder:"
 echo "$REMINDER"
 
+# Insert the reminder as a plist file
 cat > ~/Library/LaunchAgents/com.approductive.remindersapp.$TIMESTAMP.plist <<EOF
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
@@ -59,5 +88,9 @@ cat > ~/Library/LaunchAgents/com.approductive.remindersapp.$TIMESTAMP.plist <<EO
 </dict>
 </plist>
 EOF
+
+# Set permissions, then load into launchd using launchctl
 chmod 644 ~/Library/LaunchAgents/com.approductive.remindersapp.$TIMESTAMP.plist
 launchctl load ~/Library/LaunchAgents/com.approductive.remindersapp.$TIMESTAMP.plist
+
+### Create Reminder Ends ###
